# app/Docs/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FileHosting API",
        "description": "API позволяет загружать файлы на сервер, получать их метаданные и содержимое после загрузки, а также изменять некоторые метаданные загруженного файла.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://file/api/",
            "description": "Основной сервер FileHosting API"
        }
    ],
    "paths": {
        "/auth/files/{id}": {
            "post": {
                "summary": "Аутентифицироваться и авторизоваться.",
                "operationId": "auth",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/fileId"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "description": "Если viewingStatus = private и нужно дать только права на чтение, то ввести visibilityPassword; дать права на чтение и запись - modifyPassword.",
                                        "type": "string",
                                        "maxLength": 22,
                                        "minLength": 8,
                                        "pattern": "[a-zA-Z0-9!@#$%\\^&*\\(\\)\\-—_+=;:,\\.\\/?\\\\|`~\\[\\]{}]+"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пользователь аутентифицирован и авторизован.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "$ref": "#/components/schemas/Status"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationErrorResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UserIsNotAuthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Status": {
                "description": "Статус выполнения запроса.",
                "type": "string",
                "enum": [
                    "success",
                    "fail",
                    "error"
                ]
            },
...
        },
        "responses": {
            "UserIsNotAuthorized": {
                "description": "Пользователь не аутентифицирован и не авторизован.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "$ref": "#/components/schemas/Status"
                                },
                                "data": {
                                    "description": "Полезная нагрузка ответа.",
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке аутентификации и авторизации.",
                                            "type": "string"
                                        },
                                        "links": {
                                            "description": "HATEOAS",
                                            "properties": {
                                                "auth": {
                                                    "description": "URL, чтобы аутентифицироваться и авторизоваться.",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ValidationErrorResponse": {
                "description": "Ошибки валидации входных данных.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "errors": {
                                    "description": "Ошибки валидации.",
                                    "properties": {
                                        "Название поля, не прошедшее проверку валидации.": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
            "fileId": {
                "name": "id",
                "in": "path",
                "description": "ID загруженного файла.",
                "required": true,
                "allowEmptyValue": false,
                "schema": {
                    "type": "integer"
                }
            }
        }
    }
}